# swagger.yml
openapi: 3.1.0
info:
  title: CatMapper API
  version: 1.1.0
  description: |
    Public and authenticated endpoints for CatMapper (SocioMap and ArchaMap).

    Canonical REST endpoints are now exposed under `/api`. Existing unprefixed
    routes remain available as deprecated compatibility aliases during the
    migration.

    Notes on auth:
    - Most read endpoints do not require authentication.
    - Write/task endpoints support either:
      1) `Authorization: Bearer <token>`
      2) `X-API-Key: <key>` (or `Authorization: ApiKey <key>`)
      3) Backward-compatible body credentials in `cred` (`{ userid, key }`).
servers:
  - url: https://api.catmapper.org/api
    description: Production canonical REST API
  - url: http://localhost:5000/api
    description: Local development canonical REST API
...
paths:
...
  /databases/{database}/metadata/domains:
    get:
      tags: [Metadata]
      summary: List public domains
      description: 'Canonical REST endpoint. Deprecated compatibility alias: `/metadata/domains/{database}`.'
      parameters:
        - $ref: '#/components/parameters/databasePath'
      responses:
        '200':
          description: Domain rows
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '500':
          $ref: '#/components/responses/InternalError'
...
  /allDatasets:
    get:
      tags: [Metadata]
      summary: List all datasets in a database
      parameters:
        - $ref: '#/components/parameters/databaseQuery'
      responses:
        '200':
          description: Dataset list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
        '500':
          $ref: '#/components/responses/InternalError'
  /datasetDomains:
    post:
      tags: [Metadata]
      summary: Domains represented in dataset ties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [database, cmid]
              properties:
                database:
                  type: string
                cmid:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                children:
                  oneOf:
                    - type: boolean
                    - type: string
      responses:
        '200':
          description: Domain labels
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '500':
          $ref: '#/components/responses/InternalError'
  /metadata/domains/{database}:
    get:
      tags: [Metadata]
      summary: Public domains
      parameters:
        - $ref: '#/components/parameters/databasePath'
      responses:
        '200':
          description: Domain list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /metadata/subdomains/{database}:
    get:
      tags: [Metadata]
      summary: Public subdomains
      parameters:
        - $ref: '#/components/parameters/databasePath'
      responses:
        '200':
          description: Subdomain list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /metadata/properties/{database}:
    get:
      tags: [Metadata]
      summary: PROPERTY-node property table
      description: |
        Public endpoint that returns a tabular export of all key/value properties
        found on all `PROPERTY` nodes for the selected database.
        Supported databases are `sociomap` and `archamap`.
      parameters:
        - name: database
          in: path
          required: true
          schema:
            type: string
            enum: [sociomap, archamap]
          description: Target database to query.
      responses:
        '200':
          description: PROPERTY node property table
          content:
            application/json:
              schema:
                type: object
                properties:
                  database:
                    type: string
                    enum: [sociomap, archamap]
                  table:
                    type: array
                    items:
                      type: object
                      properties:
                        nodeID:
                          type: string
                        CMName:
                          type: string
                        property:
                          type: string
                        value: {}
                      required: [nodeID, CMName, property]
                      additionalProperties: false
                required: [database, table]
                additionalProperties: false
        '400':
          description: Invalid database name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    databaseQuery:
      name: database
      in: query
      required: true
      schema:
        type: string
      description: Database identifier (for example `SocioMap` or `ArchaMap`).
    databasePath:
      name: database
      in: path
      required: true
      schema:
        type: string
      description: Database identifier (for example `SocioMap` or `ArchaMap`).
...
  responses:
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        error_details:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
                nullable: true
              field:
                type: string
              code:
                type: string
              message:
                type: string
            additionalProperties: true
      additionalProperties: true
...